+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2020-02-17
+## Added
+- Params argument for validations
+
+## Changed
+- Bumped version.
+- Validation templates are now '.erb' instead of '.rb'.
+- Updated validation USAGE.
+
 ## [0.4.0] - 2020-02-16
 ### Added
 - Generator for dry-monad.